V1
Finalized V1 : exportSavesToWiiU.bat : OK

The FTP server has an issue : it can't handle files timestamps (dates). I've tried using WinSCP and FileZilla, same date = 01/01/1970

Injecting "simple" saves (such as "Fast Racing Neo" one) will work but for some games (BOTW, WWHD, MK8...) saves will be invalidated by the Wii-U because of incoherents timestamps (dates) on files. Leading in starting a new save or hanging on the splash screen (MK8)

The workaround consists in forcing the timestamp to 0 (=01/01/200) : files are injected with this date (0) and so as they are not anteriors to the timestamp value (0) and will not be rejected.